Proyecto/cv-project/backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
stripe==6.7.0
python-nmap==0.7.1
Werkzeug==2.3.7
psycopg2-binary==2.9.7
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
